Renamed 'DocumentModel' to 'ThinkyDocumentModel'
